chore: Add @smastrom/react-rating package and update dependencies
This commit adds the "@smastrom/react-rating" package to the project and updates the package.json and package-lock.json files to include the new dependency. The "@smastrom/react-rating" package is added to the frontend package.json file with version "^1.5.0" and is also included in the package-lock.json file. Additionally, the peer dependencies for "@smastrom/react-rating" are specified as "react" and "react-dom" with a minimum version of "18".