feat: Add date-fns library for date manipulation
This commit adds the date-fns library to the project, specifically version 3.6.0. The library provides a set of functions for manipulating and formatting dates, which will be used to handle the last updated date in the books data table. This enhancement improves the accuracy and consistency of date operations in the application.