feat: Update addBookFormSchema to allow optional imageUrl
This commit updates the addBookFormSchema in the add-book-form.schema.ts file to allow the imageUrl field to be optional. Previously, the field was required, but now it can be left empty if the user does not have an image URL for the book. This change improves the flexibility and usability of the add book form.