feat: Allow optional imageUrl in CreateBookDto
This commit updates the CreateBookDto in the create-book.dto.ts file to allow the imageUrl field to be optional. Previously, the field was required, but now it can be left empty if the user does not have an image URL for the book. This change improves the flexibility and usability of the CreateBookDto.