feat: Add book creation with associated shelf and user
This commit adds the ability to create a new book with an associated shelf and user. The `createNewBook` method in the `BooksService` now accepts a `CreateBookDto` object that includes the `userId` and `shelfId` properties. The book is created with the provided data, and the user and shelf are connected to the book using the `connect` and `connectOrCreate` options in the Prisma query. This feature enhances the functionality of the application by allowing users to associate books with specific shelves and users when creating new books.