feat: Include shelf information when retrieving user books
This commit modifies the `findAllBooksByUserId` method in the `BooksService` to include the `shelf` information when retrieving user books. By adding the `include: { shelf: true }` option to the Prisma query, the service now returns the associated shelf for each book. This enhancement improves the functionality of the application by providing users with the ability to view the shelf information along with their books.