feat: Add MY_BOOKSHELF constant for default shelfId in AddBookForm
This commit adds the MY_BOOKSHELF constant to the constants file in the lib directory. The MY_BOOKSHELF constant is used as the default shelfId when adding a new book in the AddBookForm component. This change enhances the functionality of the application by providing a default shelf for new books.