feat: Include book property in Shelf type
This commit adds the `book` property to the `Shelf` type in the `shelf.type.ts` file. The `book` property is an array of `Book` objects and represents the books associated with the shelf. This change enhances the functionality of the application by providing a way to store and retrieve book information for each shelf.