feat: Add default colour to shelves and create unique index for userId and colour
This commit adds a default colour to shelves by modifying the "Shelf" table in the backend/prisma/schema.prisma file. The "colour" column is added with a default value of '#ffffff'. Additionally, a unique index is created for the combination of "userId" and "colour" columns in the "Shelf" table. This change enhances the functionality of the application by providing a default colour for shelves and ensuring that each user can have only one shelf with a specific colour.